updated models.py
